Remove all aliases to get/list/create/set/delete commands
This PR removes the `ocdev component X` commands for the following
commands:

```sh
Available Commands:
  application Perform application operations
  catalog     Catalog related operations
  completion  Output shell completion code
  create      Create new component
  delete      Delete existing component
  get         Get currently active component
  help        Help about any command
  list        List all components in the current application
  project     Perform project operations
  push        Push source code to component
  set         Set active component.
  storage     Perform storage operations
  url         Expose component to the outside world
  version     Print the version of ocdev
```
